pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Gabriel Alves", email = "gabriel52.alves@gmail.com"},
]
dependencies = [
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=75.1.0",
    "django>=5.1.1",
    "Pillow>=10.4.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "drf-spectacular>=0.27.2",
    "python-dotenv>=1.0.1",
    "netifaces>=0.11.0",
    "django-filter>=24.3",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"